frameworks: base: PhoneWindow: Let width always match parent for multi-window
For multiwindow, the window's width need be always controlled by bounds
in activity stack, so phone window's width need always match parent --
will match the bounds of activity stack.

Also fix the bug for isMWPanel: 0, -1, -2 are all valid stack id which
are for single window; so for multi-window, its stack id must be larger
than 0.

Also DocumentsUI's theme need be changed to multi-window version, or it
will generate a smaller window on the parent. The floating style has
effect with this issue which need be fixed in future.

Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>
Reviewed-by: ladehunter <ladehunter@gmail.com>