frameworks: base: Fix sereral shadow related issues
For WeChat, shadow is too wide to let the content window much thinner
than before, which causes the buttons in Welcome activity are overlapped.
So let phone window intializing width wider.

For QQ, change the login activity background color whiter to fit QQ real
background.

For WanDouJia, fix crash issue: its root view is not ViewGroup, but
originally, we always treat it as view group. So let the view be in
parent content viewgroup to fix this issue.

Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>
Reviewed-by: Li Bing <libing@emindsoft.com.cn>