frameworks: base: Animation: Remove redundant animation
Firstly, update mWinFrame in ViewRootImpl in time to avoid the incorrect
animation.

After the fix, the animation is still redundant, so for multi-window
application, always use current frame for animating to freeze the
animation (not change position), so user can not feel animating.

After bypass the redundant animation, some of application windows may
freeze about 0.5 second when toggle full screen to wait their animating
end, but the user rarely can feel, notice, or care about it.

Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>
Reviewed-by: Li Bing <libing@emindsoft.com.cn>