frameworks: base: VLC: Fix issue for no response when click VLC SurfaceView
Normally, when click SurfaceView, the player will toggle control gui,
VLC uses onTouch in Activity to perform this action.

But in AOSP-5.1, when user click SurfaceView, the event is eaten by
views, then onTouch in Activity can not get the event, so cause this
issue.

After continuing to route the event by force for VLC case, this issue
can be solved.

Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>
Reviewed-by: Li Bing <libing@emindsoft.com.cn>