frameworks: base: LineRect: Fix one issue and add one feature
When hold down the application titlebar and drag to the left or right
side of screen, the screen will appear a region of half screen area with
translucent background.

When hold down the edge of the application and drag to change its' size,
we still use original implementation -- the animation and transluncent
background have no effect.

Signed-off-by: wangzhixu <potatomagic@163.com>
Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>