frameworks: base: WeChat: Let its all popup window show correctly
For popup window height, it is calculated in app side, which may be
truncated, so we have to check it in PopupWindow in APP side.

For popup window x position, WeChat will let it must be within the APP
window, but its calculation assumes left-top must be 0, which isn't true
in OPENTHOS. So we need use the last mouse position to re-calculate it.

Also clean up the related code which we can see.

Signed-off-by: Chen Gang <chengang@emindsoft.com.cn>
Reviewed-by: Chen Peng <0123hhtx@163.com>