frameworks: base: Fix four issues
Set the font size to the maximum, the start menu sorting method is incomplete

Set the font size to the maximum, and the input method does not show one line

Set the system language to English (United State). After restart, the right language of the taskbar is still display as Chinese

After switching the language to English, when restart the system, the taskbar, battery or calendar display as chinese.

Signed-off-by: Luo Junhuan <forward_m@126.com>
Signed-off-by: Wang Zhixu <potatomagic@163.com>